Adicionando comando unknown
