--- commands/start.py
@@ -1,5 +1,0 @@
-def start(update, context):
-    context.bot.send_message(
-        chat_id=update.effective_chat.id,
-        text="I'm Wrench Jr."
-    )
