+++ commands/unknown/command.py
@@ -2,13 +2,6 @@ import logging
 
 from telegram.ext import Filters, MessageHandler
 
-from core import TelegramCore
-
-# Enable logging
-logging.basicConfig(
-    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s", level=logging.INFO
-)
-
 logger = logging.getLogger(__name__)
 
 def unknown(update, context):
@@ -17,5 +10,4 @@ def unknown(update, context):
         text="Sorry, I didn't understand that command.",
     )
 
-def config_handlers(instance: TelegramCore):
-    instance.add_handler(MessageHandler(Filters.command, unknown))
+unknown_handler = MessageHandler(Filters.command, unknown)
